requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.0.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
evdev==0.7.0
feedparser==5.2.1
Flask==1.0.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2-binary==2.7.5
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.5.3
pylint==2.0.1
pymacaroons==0.13.0
pymemcache==1.4.4
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.2
python-debian==0.1.32
pytz==2018.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.14.1
wrapt==1.10.11
xkit==0.0.0
zope.interface==4.3.2
